+++ CHANGELOG.md
@@ -58,6 +58,12 @@
 - Add a hinge-angle-based bending force to `wp.sim.VBDIntegrator`.
 - Add an example to show mesh sampling using a CDF
   ([GH-476](https://github.com/NVIDIA/warp/issues/476)).
+- Support Python/SASS correlation in Nsight Compute reports by emitting `#line` directives in CUDA-C code.
+  This setting is controlled by `wp.config.line_directives` and is `True` by default.
+  ([docs](https://nvidia.github.io/warp/profiling.html#nsight-compute-profiling),
+   [GH-437](https://github.com/NVIDIA/warp/issues/437))
+- Add example of a distributed Jacobi solver using `mpi4py` in `warp/examples/distributed/example_jacobi_mpi.py`
+  ([GH-475](https://github.com/NVIDIA/warp/issues/475)).
 
 ### Changed